# users/openapi.yaml
openapi: 3.0.0
info:
  title: Assist Health Users API
  version: 1.0.0
  description: API specification for Assist Health user management system
  contact:
    name: API Support
    email: support@assisthealth.com
servers:
  - url: '{baseUrl}/api/v1'
    variables:
      baseUrl:
        default: http://localhost:3000
        description: Base URL of the API server
components:
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: Error message describing what went wrong
        data:
          type: object
          nullable: true
          description: Additional error details if any
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
          description: Success message describing the operation result
        data:
          type: object
          description: Response data
...
    Package:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        description:
          type: string
        features:
          type: array
          items:
            type: string
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /packages/{packageId}:
    get:
      summary: Get package by ID
      operationId: getPackageById
      parameters:
        - name: packageId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Package retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Package'
        '404':
          $ref: '#/components/responses/NotFound'
...